restapi/user/urls.py
from django.urls import path
from rest_framework_simplejwt import views as jwt_views
from .views import HomePage, Register, Logout, activate

urlpatterns = [
    path('token/', jwt_views.TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', jwt_views.TokenRefreshView.as_view(), name='token_refresh'),
    path('register/', Register.as_view(), name='register'),
    path('logout/', Logout.as_view(), name='logout'),
    path('activate/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/',  
        activate.as_view(), name='activate'),
    path('home/', HomePage.as_view(), name='homepage')
]