requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
django-cors-headers==4.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2